+++ actix-files/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Opt-In filesize threshold for faster synchronus reads that allow for 20x better performance.
 - Minimum supported Rust version (MSRV) is now 1.75.
 
 ## 0.6.6